prepared the necessary environment for utilizing the database in the project by adding the required files and implementing some modifications.
